Run Bouncer_CDN job in core hours
This is in response to an incident where we get called up at 4am if
there is a problem with the Bouncer_CDN. In the incident review [1] we
didn't think there was any particular reason why this was run out of
hours (it was previously triggered off a job that runs at 3/4am [2]).

The previous commit removed the downstream triggering of this job, so
this commit sets the time to run ~10am in the morning so that only
problems with it can occur within core GDS hours.

[1]: https://docs.google.com/document/d/1ZqCHkf2XPfBoQUF3K6SLnlBO94TruABGt05HFWYtF1A/edit
[2]: https://github.com/alphagov/govuk-puppet/blob/7a85a1d568300bafc0ea3eb0d6681e466268ba19/modules/govuk_jenkins/templates/jobs/transition_load_all_data.yaml.erb#L25-L29