Only set cors headers for fonts
This is a quick demo of a regex I copy and pasted that might be enough to only return cross origin headers for font files. Unfortunately as these headers have existed for a long time I'm not sure if this would have wider consequences. I also don't know nginx location handling and precedence to know if this would stop the proxy pass location block or would work in conjunction with it - I have a feeling it might be the latter.