Add redirect for Whitehall assets on assets host
Whitehall does a rather strange where it embeds it's asset URLs into
some content items when they use a govspeak asset. This means that,
unlike other asset URLs, some of these need to be maintained
permanently. To handle maintaining these permanently this code adds a
redirect on the asset hostnames so any requests for
files to /government/assets are redirected.

I didn't think there'd be any scope for there to be requests to
draft-assets since Whitehall doesn't use that hostname, so I've only
done this for the live asset hostname.