Disable Sentry for Search API in test environment.
Search API in the test environment has been spamming Sentry and
this has been inconveniencing developers. For example it's been
continuously sending about 6 Faraday::TimeoutError per hour:
https://sentry.io/organizations/govuk/issues/1711334233/

Ideally we would just deploy the latest version of Search API into the
test environment and it'd pick up the client-side environment filtering
which went into version 2.6.0 of the govuk_app_config gem, but the
Deploy_App Jenkins job is broken for Search API in the test environment
because of a Ruby version issue (the job fails with `rbenv: version
`2.7.2' is not installed`).

This assumes that setting the SENTRY_DSN environment variable to the
empty string results in disabling the Sentry client without preventing
the app from starting. Even if that's not the case and it breaks Search
API in the test environment, that's not really a problem as we're not
using it.