Add Ngnix cache for Mapit
This enables a nginx cache for Mapit. This is a temporary fix, whilst
work is being done to improve caching within the Mapit application
itself. Currently Mapit doesn't cache any json request, this causes every
request to require a database look up, which is expensive and adds
latency. The Mapit data is infrequently updated so makes it a good
candidate for caching responses.