Fetch a top-level "message" key from JSON healthcheck response
Some things which were in the "checks" object are now moving to their
own endpoints, separate from the main app healthcheck.  This change is
so that these new endpoints can return just

    { "status": "ok", "message": "details" }

Rather than

    { "status": "ok", "checks":
      { "checkname": { "status": "ok", "message": "details" } }
    }

In particular, right now the new signon expiring API tokens check is
not reporting *which* tokens are expiring, because it is returning
that information in a top-level "message" key.